상품 응답 객체에 판매자 정보 출력

+++ server/wonprice/src/main/java/main/wonprice/domain/product/service/ProductService.java
@@ -36,6 +36,6 @@ public interface ProductService {
 
     Page<Product> searchProductsByTitle(String keyword, Pageable pageable);
 
-    Integer getProductWishCount(Long productId);
+    Long getProductWishCount(Long productId);
 }
